+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changelog
 ### Version 1.0.6
 * Resolved issue with Angle still affecting the emitter despite being deactivated. Thanks ludek17052.
+* The position of a newly loaded effect is now set to the position of the old one. Thanks ChinQingl.
 ### Version 1.0.5
 * Fix export file dialog defaulting to an incorrect path
 * Fixed image name error when loading old particle effects with whole paths listed in the particle file